Move ResultReporter to nunit.common
