+++ src/NUnitConsole/nunit4-console/TestEventHandler.cs
@@ -1,12 +1,10 @@
 // Copyright (c) Charlie Poole, Rob Prouse and Contributors. MIT License - see LICENSE.txt
 
-using System;
 using System.Xml;
 using NUnit.Common;
 using NUnit.Engine;
-using NUnit.TextDisplay;
 
-namespace NUnit.ConsoleRunner
+namespace NUnit.TextDisplay
 {
     /// <summary>
     /// TestEventHandler processes events from the running
