+++ src/NUnitConsole/nunit4-netcore-console/nunit4-netcore-console.csproj
@@ -50,7 +50,6 @@
     <Compile Include="..\nunit4-console\Options\OutputSpecification.cs" Link="Options\OutputSpecification.cs" />
     <Compile Include="..\nunit4-console\Options\TestNameParser.cs" Link="Options\TestNameParser.cs" />
     <Compile Include="..\nunit4-console\RequiredExtensionException.cs" Link="RequiredExtensionException.cs" />
-    <Compile Include="..\nunit4-console\ResultReporter.cs" Link="ResultReporter.cs" />
     <Compile Include="..\nunit4-console\TestEventHandler.cs" Link="TestEventHandler.cs" />
     <Compile Include="..\nunit4-console\Utilities\SaveConsoleOutput.cs" Link="Utilities\SaveConsoleOutput.cs" />
     <Content Include="..\..\..\LICENSE.txt" Link="LICENSE.txt" />
